remove 'Optimization' module from 'Engineering' module
